refactors wine comment form
