--- src/components/NewCommentForm.js
@@ -1,55 +1,0 @@
-// // Setup this form component but left commented out. May be useful but will need to return to this later once had a look at in more depth
-// const Comment = ({ message }) => {
-//   return (
-//     <>
-//       <h1>New Comment</h1>
-//       <p>{message.text}</p>
-//       <p>{message.user}</p>
-//     </>
-//   );
-// };
-
-// export default Comment;
-
-import React, { useState } from "react";
-import Button from "react-bootstrap/Button";
-import Modal from "react-bootstrap/Modal";
-
-function Comment() {
-  const [show, setShow] = useState(false);
-
-  const handleClose = () => setShow(false);
-  const handleShow = () => setShow(true);
-
-  return (
-    <>
-      <Button variant="primary" onClick={handleShow}>
-        Launch static backdrop modal
-      </Button>
-
-      <Modal
-        show={show}
-        onHide={handleClose}
-        backdrop="static"
-        keyboard={false}
-      >
-        <Modal.Header closeButton>
-          <Modal.Title>Modal title</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          I will not close if you click outside me. Don't even try to press
-          escape key.
-        </Modal.Body>
-        <Modal.Footer>
-          <Button variant="secondary" onClick={handleClose}>
-            Close
-          </Button>
-          <Button variant="primary">Understood</Button>
-        </Modal.Footer>
-      </Modal>
-    </>
-  );
-}
-
-export default <Comment />;
-// render(<Comment />);
